requirements/tests.txt
-r base.txt

coverage==3.7
django-factory_boy==0.1.6